+++ tests/testthat/test_matchSpectra.R
@@ -4,7 +4,6 @@ test_that("CompareSpectraParam works", {
 
     expect_error(CompareSpectraParam(tolerance = 1:3), "positive number")
     expect_error(CompareSpectraParam(ppm = -4), "positive number")
-    expect_error(CompareSpectraParam(threshold = 1:2), "length 1")
 
     res <- CompareSpectraParam(other_param = 5, b = 3)
     expect_equal(res@dots, list(other_param = 5, b = 3))
@@ -14,6 +13,27 @@ test_that("CompareSpectraParam works", {
     expect_equal(res$other_param, 5L)
     expect_equal(res$b, 3L)
     expect_equal(res$ppm, 5)
+})
+
+test_that(".valid_threshfun works", {
+    res <- .valid_threshfun(function(x) which(x > 0.4))
+    expect_true(length(res) == 0)
+    res <- .valid_threshfun(function(x) x > 0.4)
+    expect_true(length(res) == 0)
+    res <- .valid_threshfun(NULL)
+    expect_true(length(res) == 0)
+
+    res <- .valid_threshfun("f")
+    expect_match(res, "function")
+
+    res <- .valid_threshfun(function(z) z / 3)
+    expect_match(res, "integer")
+
+    res <- .valid_threshfun(function(z) -1)
+    expect_match(res, "integer")
+
+    res <- .valid_threshfun(function(z) TRUE)
+    expect_match(res, "logical")
 })
 
 test_that(".compare_spectra, .compare_spectra_without_precursor work", {
@@ -48,7 +68,7 @@ test_that(".compare_spectra, .compare_spectra_without_precursor work", {
 test_that(".get_matches_spectra, matchSpectra,CompareSpectraParam works", {
     csp <- CompareSpectraParam(
         requirePrecursor = FALSE,
-        THRESHFUN = function(x) x == apply(x, 1, max, na.rm = TRUE))
+        THRESHFUN = function(x) which.max(x))
     res <- matchSpectra(pest_ms2, minimb, csp)
     expect_equal(res@matches$query_idx, 1:13)
 
@@ -90,7 +110,8 @@ test_that("matchSpectra,MatchForwardReverseParam works", {
     res <- matchSpectra(pest_ms2, minimb, mp)
     expect_equal(res@matches$query_idx, 1:13)
     expect_equal(colnames(res@matches), c("query_idx", "target_idx", "score",
-                                          "reverse_score", "presence_ratio"))
+                                          "reverse_score", "presence_ratio",
+                                          "matched_peaks_count"))
 
     mp <- MatchForwardReverseParam(requirePrecursor = TRUE,
                                    THRESHFUN = function(x) which.max(x))
@@ -102,4 +123,10 @@ test_that("matchSpectra,MatchForwardReverseParam works", {
     mp <- MatchForwardReverseParam(requirePrecursor = TRUE)
     res <- matchSpectra(pest_ms2, minimb, mp)
     expect_equal(unique(res@matches$query_idx), c(2, 4, 6, 8))
+
+    mp <- MatchForwardReverseParam(requirePrecursor = TRUE,
+                                   THRESHFUN_REVERSE = function(z) z > 0.9)
+    res_2 <- matchSpectra(pest_ms2, minimb, mp)
+    expect_true(all(res_2@matches$reverse_score > 0.9))
+    expect_equal(unique(res_2@matches$query_idx), c(2, 4, 8))
 })